requirements.txt
pandas~=2.2.3
matplotlib~=3.10.3
XlsxWriter~=3.2.3
PyYAML~=6.0.2
natsort~=8.4.0